requirements.txt
google-cloud-firestore==2.0.1
getmac
sty